refactor. 데이터를 page폴더에서 처리 후 하위 컴포넌트로 내려주었다.
